Added plists (BranchList, ServiceHours and Holidays) and branch images to project.
BranchList.plist contains the list of branches and their pertinent info.
ServiceHours.plist contains the list of schedules for various branches.
Holidays.plist contains the list of holidays, when the Library will be closed.
All branch images are in PNG format and are 233 px by 175px, with a reflection (to give it more "zazz").

Signed-off-by: Rob Dumas <rob@bitsandbooks.com>